Adjages recent julia and MelGeneralizedCepstrums changes
now requires julia v0.4

+++ test/dtw.jl
@@ -6,9 +6,9 @@ println("testing: Dynamic time warping")
 
 let
     # template sequence [1,2,3,4]
-    v1 = float64([1 2 3;1 2 4;1 8 5;10 3 6])'
+    v1 = [1. 2 3;1 2 4;1 8 5;10 3 6]'
     # sequence to be aligned [1 2 2 3 4]
-    v2 = float64([1 2 3;1 2 4;1 2 5; 1 8 5;10 3 6])'
+    v2 = [1. 2 3;1 2 4;1 2 5; 1 8 5;10 3 6]'
 
     # Perform alignment
     d = DTW(bstep=1, fstep=0)
@@ -19,8 +19,8 @@ let
 end
 
 let
-    v1 = float64([0; 1; 2; 3; 4; 5])'
-    v2 = float64([0; 0; 1; 2; 3; 4; 4; 5;])'
+    v1 = [0.; 1; 2; 3; 4; 5]'
+    v2 = [0.; 0; 1; 2; 3; 4; 4; 5;]'
 
     # Perform alignment
     d = DTW(bstep=1, fstep=0)
